+++ packages/frontend/src/scripts/get-dom-node-or-null.ts
